+++ src/lib.rs
@@ -482,8 +482,8 @@ pub(crate) mod symbol_table;
 pub(crate) mod type_builders;
 pub(crate) mod vl_data;
 pub(crate) mod width;
-#[cfg(feature = "zfp")]
-pub(crate) mod zfp;
+#[cfg(all(test, feature = "zfp", feature = "std"))]
+mod zfp_crosscheck;
 
 #[cfg(feature = "provenance")]
 pub(crate) mod provenance;